Update explainer for the new spec
This updates the explainer text according to the new spec we agreed in
the 09-15-2020 CG meeting. Some of the text is taken from [the first
version](https://github.com/WebAssembly/exception-handling/blob/master/proposals/old/Exceptions-v1.md)
of the proposal. I renamed `catch_br` to `delegate` based on the
perceived consensus in #133.

There are some minor issues that need clarification, some of which is
currently being discussed in the repo:
- `rethrow`'s immmediate (or label) argument. I made it match the status
  in the first proposal for now. I believe keeping it makes it more
  versatile and lets us handle more languages.
- It is not clear whether `rethrow` and `delegate`'s label (=immediate)
  argument should count non-try block-like structures, such as blocks
  and loops. It is currently being discussed for `delegate`.
- `delegate`'s opcode should be decided. We exhausted all control-flow
  opcodes that begin with `0x0`.