Add detailed example for rethrow to explainer
The current explainer has a detailed example on `delegate`, which was
adapted from
https://github.com/WebAssembly/exception-handling/issues/146#issue-794482909,
but not the one for `rethrow` from the same issue
(https://github.com/WebAssembly/exception-handling/issues/146#issuecomment-777714491).
This adds the detailed `rethrow` example to the explainer, with the
some cosmetic changes to the label names and comments.