Add exceptional return to func_invoke in embedding doc
This is a basic start of what we'll need to add to the embedder spec.
I'm not sure if we also need a way to allow exceptions to bubble
from the embedder into wasm ; I don't see any reference to what
happens when the instance wants to call an import, so maybe not.
Also the references aren't quite right, but I may need help
to fix that