[js-api] Update wasm-module-builder.js for exnref
This pulls in some exnref-related changes from
https://github.com/v8/v8/blob/main/test/mjsunit/wasm/wasm-module-builder.js
but does not sync with it completely. That v8 version contains a lot of
features from new proposals that are not really relevant to EH and
exnref, so this pulls only the relevant EH parts in.

This also renames `kWasmStmt` to `kWasmVoid` as in the V8 version,
because it looks more intuitive. Also renames `kTagAttribute` to
`kExceptionAttribute`.