[js-api] Re-add old tests to test/js-api/exception
These two tests used to be in
https://github.com/WebAssembly/exception-handling/tree/main/test/js-api/exception
and were moved into
https://github.com/WebAssembly/exception-handling/tree/main/test/legacy/exceptions/js-api
in #305.

I'm planning new version of these tests that use `try_table` and
`throw_ref` in
https://github.com/WebAssembly/exception-handling/tree/main/test/js-api/exception,
but it wouldn't require to rewrite the whole tests, but copying these
tests into the this directory and modify them in a single PR makes
Github think these are brand-new files, resulting in a large diff
containing the whole files that is difficult to review (and which has
been reviewed and in the repo for a long time already).

So I'm making a PR that only re-adds these file here so that I can add
changes to these files in another PR.