Add link to legacy proposal in explainer
I found that the legacy and pre-legacy proposals have links to their
immediate predecessor and successor but the current proposal does not
have a link to its (immediate) predecessor. Given that the current
proposal is not yet supported without flags in many browsers, it would
help people who want to look at the legacy proposal but don't know there
is a separate `legacy/` directory.